+++ exec.go
@@ -3,9 +3,12 @@ package bob
 import (
 	"context"
 	"database/sql"
+	"errors"
 
 	"github.com/stephenafamo/scan"
 )
+
+var ErrHookableTypeMismatch = errors.New("hookable type mismatch: the slice type is not hookable, but the single type is")
 
 type (
 	MapperModder interface {
@@ -115,25 +118,44 @@ func One[T any](ctx context.Context, exec Executor, q Query, m scan.Mapper[T]) (
 }
 
 func All[T any](ctx context.Context, exec Executor, q Query, m scan.Mapper[T]) ([]T, error) {
-	return Allx[T, []T](ctx, exec, q, m)
+	return Allx[SliceTransformer[T, []T]](ctx, exec, q, m)
 }
 
-// Allx takes 2 type parameters. The second is a special return type of the returned slice
-// this is especially useful for when the the [Query] is [Loadable] and the loader depends on the
-// return value implementing an interface
-func Allx[T any, Ts ~[]T](ctx context.Context, exec Executor, q Query, m scan.Mapper[T]) (Ts, error) {
+// SliceTransformer is a [Transformer] that transforms a scanned slice of type T into a slice of type Ts.
+type SliceTransformer[T any, Ts ~[]T] struct{}
+
+func (SliceTransformer[T, Ts]) TransformScanned(scanned []T) (Ts, error) {
+	return Ts(scanned), nil
+}
+
+type Transformer[T any, V any] interface {
+	TransformScanned([]T) (V, error)
+}
+
+// Allx in addition to the [scan.Mapper], Allx takes a [Transformer] that will transform the scanned slice into a different type.
+// For common use cases, you can use [SliceTransformer] to transform a scanned slice of type T into a custom slice type like ~[]T.
+func Allx[Tr Transformer[T, V], T any, V any](ctx context.Context, exec Executor, q Query, m scan.Mapper[T]) (V, error) {
+	var typedSlice V
 	var err error
+
+	_, isTransformedHookable := any(typedSlice).(HookableType)
+	_, isSingleHookable := any(*new(T)).(HookableType)
+	// If the transformed type is not hookable, but the single type is,
+	// return an error
+	if !isTransformedHookable && isSingleHookable {
+		return typedSlice, ErrHookableTypeMismatch
+	}
 
 	if h, ok := q.(HookableQuery); ok {
 		ctx, err = h.RunHooks(ctx, exec)
 		if err != nil {
-			return nil, err
+			return typedSlice, err
 		}
 	}
 
 	sql, args, err := Build(ctx, q)
 	if err != nil {
-		return nil, err
+		return typedSlice, err
 	}
 
 	if l, ok := q.(MapperModder); ok {
@@ -144,10 +166,14 @@ func Allx[T any, Ts ~[]T](ctx context.Context, exec Executor, q Query, m scan.Ma
 
 	rawSlice, err := scan.All(ctx, exec, m, sql, args...)
 	if err != nil {
-		return nil, err
+		return typedSlice, err
 	}
 
-	typedSlice := Ts(rawSlice)
+	var transformer Tr
+	typedSlice, err = transformer.TransformScanned(rawSlice)
+	if err != nil {
+		return typedSlice, err
+	}
 
 	if l, ok := q.(Loadable); ok {
 		for _, loader := range l.GetLoaders() {
@@ -160,12 +186,6 @@ func Allx[T any, Ts ~[]T](ctx context.Context, exec Executor, q Query, m scan.Ma
 	if h, ok := any(typedSlice).(HookableType); ok {
 		if err = h.AfterQueryHook(ctx, exec, q.Type()); err != nil {
 			return typedSlice, err
-		}
-	} else if _, ok := any(*new(T)).(HookableType); ok {
-		for _, t := range typedSlice {
-			if err = any(t).(HookableType).AfterQueryHook(ctx, exec, q.Type()); err != nil {
-				return typedSlice, err
-			}
 		}
 	}