+++ CHANGELOG.md
@@ -12,10 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `As` method to `clause.TableRef`, which sets the alias and return a copy of the struct. (thanks @Nitjsefni7)
 - Added the `type_system` configuration option to determine how to generate null and optional values in the generated code.
   Possible options: `github.com/aarondl/opt`, `github.com/aarondl/opt/null` or `database/sql`. The default value is `github.com/aarondl/opt`.
+- When generating code for queries, columns with dots (`.`) in their names are now assumed to be nested fields. The `All` method of the generated query will return a struct with nested fields instead of a flat struct.
+- Implement `--prefix` annotation in queries for `bobgen-psql`.
 
 ### Changed
 
 - `Mod` is now a separate field in `orm.ModQuery` and `orm.ModExecQuery`.
+- `Allx` now takes a `Transformer` type parameter to transform the result of the query.
 
 ### Removed